backend/requirements.txt
# Core API framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# LangGraph and LangChain dependencies
langgraph>=0.0.55
langchain-core>=0.1.45
langchain-openai>=0.1.0

# HTTP client for API calls
aiohttp>=3.9.0
httpx>=0.25.0

# Data validation and serialization
pydantic>=2.5.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0

# Logging and monitoring
structlog>=23.2.0

# Environment and configuration
python-dotenv>=1.0.0

# Optional: Database support (uncomment if needed)
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Optional: Vector database for enhanced KB search (uncomment if needed)  
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
